+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
+## Next Release
+* FEATURE: Ability to use just total and update without passing an iterable
+
+## 1.4.1 (2022-03-10)
+* FIX: Progress bars on 32bit systems use the correct integer type now (`UInt` -> `UInt64`)
+
 ## 1.4.0 (2021-08-12)
 
-Changes:
-* Print to `stderr` instead of `stdout` to be in line with other libraries (Contributed by @InnovativeInventor, see PR #47)
+* CHANGE: Print to `stderr` instead of `stdout` to be in line with other libraries (Contributed by @InnovativeInventor, see PR #47)